Added shared values
We need to share values between the detection functions and the
motor functions. Because of this we need to setup shared values.
They will be kept in the motor.h header file. The var motor_status
can only be changed by the motor functions and motor_adjustment
can only be changed by the dectection functions. This why we will
not run into timing issues.