Changes to detection
First off we rename the file to a .cpp so we can use the Serial
library to debug it. Also s0 is always on the left and s2 is alway
on the right so we don't need to calculate their facing every time.
And lastly we added some debuffing code. We will remove this when
we have completed our testing.